+++ tests/core/editable/domfix2.js
@@ -1,4 +1,4 @@
-/* bender-tags: editor,unit */
+/* bender-tags: editor,unit,autoparagraphing */
 
 var doc = CKEDITOR.document;
 // This group of tests plays upon the framed content.